Version 0.21 [Ataques Enemigos, Particulas Fuego Y Hielo Implementadas, Cambios en ataque de fuego y hielo(Daño/Efecto en Area)]

+++ Assets/SCRIPTS/INTERFACES/IDamageable.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 interface IDamageable
 {
-    public void GetDamaged(float damage);
+    public void GetDamaged(int damage);
 
 
     public void OnDeath();
